Time: 0 ms (100.00%) | Memory: 7.8 MB (53.24%) - LeetSync

+++ 1817-calculate-money-in-leetcode-bank/calculate-money-in-leetcode-bank.cpp
@@ -4,12 +4,12 @@ public:
        
        int x =n/7;
        n=n%7;
-       long long monday=1 , days=1 ;
-       long long ans=0;
+       int monday=1 , days=1 ;
+       int ans=0;
        while(days<=x)
        {
-          long long end_day= monday+6;
-          long long sum = end_day*(end_day+1)/2- (monday-1)*(monday)/2;
+          int end_day= monday+6;
+          int sum = end_day*(end_day+1)/2- (monday-1)*(monday)/2;
           ans+= sum;
         monday++;
         days++;
